Refactor common search methods into a base class
